+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/NewEnchantmentsFeature.java
@@ -147,11 +147,11 @@ public class NewEnchantmentsFeature extends Feature {
 		CurseOfBloodPact.trySuckingAndRepairing(event);
 	}
 
-	@SubscribeEvent
+	@SubscribeEvent(priority = EventPriority.LOW)
 	public void onStackMaxDamage(StackMaxDamageEvent event) {
 		int lvl = event.getStack().getEnchantmentLevel(ENDURING.get());
 		if (lvl > 0)
-			event.setNewMaxDamage(event.getNewMaxDamage() + Enduring.getBonusDurability(event.getStack()) * lvl);
+			event.setNewMaxDamage(event.getNewMaxDamage() + Enduring.getBonusDurabilityPerLevel(event.getStack()) * lvl);
 	}
 
 	@SubscribeEvent
@@ -221,8 +221,9 @@ public class NewEnchantmentsFeature extends Feature {
 
 	@SubscribeEvent(priority = EventPriority.HIGH)
 	public void onBreakSpeedEnchantment(EnchantmentBonusEfficiencyEvent event) {
-		if (event.getStack().getEnchantmentLevel(HASTE.get()) > 0)
-			event.setNewEfficiency(event.getNewEfficiency() + Haste.getBonusEfficiency());
+		int hasteLvl = event.getStack().getEnchantmentLevel(HASTE.get());
+		if (hasteLvl > 0)
+			event.setNewEfficiency(event.getNewEfficiency() + Haste.getBonusEfficiency() * hasteLvl);
 		event.setNewEfficiency(event.getNewEfficiency() + Blasting.getMiningSpeedBoost(event.getStack(), event.getEntity(), event.getState()));
 		event.setNewEfficiency(event.getNewEfficiency() + Adrenaline.getMiningSpeedBoost(event.getStack(), event.getEntity(), event.getState()));
 		if (event.getStack().getEnchantmentLevel(CURSE_OF_INEFFICIENCY.get()) > 0)