+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/forging/DurabilityModifier.java
@@ -1,0 +1,7 @@
+package insane96mcp.iguanatweaksexpanded.module.mining.forging;
+
+import net.minecraft.world.item.ItemStack;
+
+public interface DurabilityModifier {
+    float getDurabilityMultiplier(ItemStack stack);
+}